Switch to a monad that allows for incremental decompression.
This allows us to decompress large files in constant space, which is a big
advantage. Unfortunately, it also comes with a fairly huge performance penalty,
at least for the moment: on my test case, we move from ~6s to ~17s. Hopefully we
can fix this in the future ... although the constant space thing is pretty
important, I'd prefer it not to cost so much.