Speed and benchmarking improvements to the core subsystems.
The primary purpose of this commit is to improve the performance of the
decompression library. In my tests, these changes improve the performance of the
library by 20-25%, depending on your particular use case. The difference should
be felt particularly well in well-compressed files, such as ASCII text.

As part of this work, I've also added a README and a benchmark for others to
use. Both should probably be updated and expanded in the future.

(This is a squashed merge of the 'speedy' branch.)