Switch out the use of the generic FingerTree library for an inlined, de-abstracted version.
This removes the external dependency, and copies the relevant source
file into the local tree. I then stripped out all of the functions we
don't use within this library, and de-abstracted from a generic measure
to one that is definitely an integer. This appears to have a fairly
significant performance benefit for the library, at the cost of not
tracking any improvements made in the upstream library.