Simplify the Adler32 logic a bit, to rid ourselves of some `fromIntegral`.
This is less of a clear win. There's some noise in the benchmarking code
that actually makes it appear that this might be a small regression, but
I think I'm going to chalk that up to -- as I said -- noise. In some
cases, and notably ones that the FingerTree inline test didn't seem to
help, this appears to gain us ~5% in performance.