Further adjustments to the Adler computation, including a case split on input size.
Most other implementations of Adler32 use different cases for blocks of
1 byte, many-but-less-than-16-or-5552 bytes, and lots of bytes. The idea
for the mid-sized chunks is that you can avoid doing a lot of
extraneous modulo operations if you batch up the adds in a way that
doesn't allow the values to roll over. In addition, we bite the bullet
and switch to unlifted word values to do the computation; this has a
number of beneficial effects, while retaining safety, including a lack
of unnecessary type conversions. Finally, shifting from `foldl` to
`foldl'` helps. Altogether, this has a very significant effect on
benchmarks that are bottlenecked by the checksum computation; our random
test cases, for example, are so bottlenecked, and see order-of-magnitude
performance improvements.