Create a custom function for recreating a blob based on our history.
Prior versions used `FingerTree` functions to first split the tree
around where we wanted to find a historical blob, and then other
`FingerTree` manipulations to reconstruct the blob as a `ByteString`.
The latter part involved quite a lot of recomputation of a new, valid
tree that we don't necessarily need in this use case, so this patch adds
a function that does away with some (although not all) of that
recreation. The performance improvement is not huge (we'd probably need
to look at redoing the entire `viewl` chain that is invoked via `rotL`
to get bigger improvements), but gets us 10% in a couple cases and
recovers some regressions in our zero tests.