Switch to using a `ByteArray#` for the `OutputWindow`, with all the ST that follows.
This patch trades the completely, 100% pure implementation of
OutputWindow to a mostly-safe, certainly-not-IO implementation
in ST. We try, as we can, to limit the safety impact of these
changes through some explicit bounds checks. The result should
be safe, and still doesn't use any direct pointer work, but has
major speed implications. Speedwise, there's still more to be
gained, here, we just need to ensure we do so without further
tweaking the safety argument.